requirements.txt
ccxt
pandas
pandas_ta
python-dotenv
requests
python-telegram-bot  # for alerts
openai